+++ frontend/lost/package.json
@@ -62,6 +62,7 @@
     "redux": "^4.0.1",
     "redux-form": "^8.2.0",
     "redux-thunk": "^2.3.0",
+    "rfdc": "^1.1.4",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^0.87.3",
     "simple-line-icons": "^2.4.1",
