+++ frontend/lost/src/components/SIA/SIA.js
@@ -126,6 +126,7 @@ class SIA extends Component {
 
   requestImageFromBackend() {
     this.props.getSiaImage(this.props.annos.image.url).then(response => {
+      // console.log("%%%%%%%%%%%%", response)
       this.setState({
         image: {
           // ...this.state.image,
@@ -138,12 +139,13 @@ class SIA extends Component {
   }
 
   render() {
+    console.log("**********", this.state.image.data)
     return (
       <div>
         {this.state.image.data ? (
           <ImageComponent
             labels={this.props.possibleLabels}
-            colors={colors}
+            // colors={colors}
             annos={this.props.annos}
             imageUrl={this.state.image.data}
             imageId={this.state.image.id}
