+++ frontend/lost/src/utils/variables.js
@@ -1,0 +1,10 @@
+let distance_between_two_vertical_points = 10
+let minimum_left_margin = 1.5
+let minimum_top_margin = 1.5
+let default_label = -1
+let default_color = 'grey'
+let colors = ['grey', "#00cc44", "#ff3300", "#0099ff", "#ff00ff", "#ffff00", "#336600", "#ff6600", "#996633"];
+export {distance_between_two_vertical_points, minimum_left_margin, minimum_top_margin, default_label, default_color, colors}
+
+
+// {"imgId":435,"imgLabelIds":[],"imgLabelChanged":false,"annotations":{"bBoxes":[],"lines":[],"points":[],"polygons":[{"labelIds":["37"],"mode":"view","status":"new","type":"polygon","data":[{"x":422.5,"y":210.5},{"x":422.5,"y":239.5},{"x":410.5,"y":225.5},{"x":410.5,"y":196.5},{"x":433.5,"y":196.5},{"x":410.5,"y":254.5}]}]},"isJunk":null}
